+++ prog/dftb+/lib_dftbplus/main.F90
@@ -70,6 +70,7 @@ module dftbp_main
   use dftbp_simplealgebra
   use dftbp_message
   use dftbp_repcont
+  use dftbp_halogenx
   use dftbp_xlbomd
   use dftbp_slakocont
   use dftbp_linkedlist
@@ -432,6 +433,12 @@ contains
       call calcDispersionEnergy(dispersion, energy%atomDisp, energy%Edisp, iAtInCentralRegion)
     end if
 
+    if (allocated(halogenXCorrection)) then
+      call halogenXCorrection%getEnergies(energy%atomHalogenX, coord, species, neighbourList,&
+          & img2CentCell)
+      energy%EHalogenX = sum(energy%atomHalogenX(iAtInCentralRegion))
+    end if
+
     call resetExternalPotentials(refExtPot, potential)
 
     if (tReadShifts) then
@@ -641,7 +648,7 @@ contains
             & tImHam.or.tSpinOrbit, tPrintMulliken, orbitalL, qBlockOut, Ef, Eband, TS, E0,&
             & extPressure, cellVol, tAtomicEnergy, tDispersion, tEField, tPeriodic, nSpin, tSpin,&
             & tSpinOrbit, tSccCalc, allocated(onSiteElements), tNegf, invLatVec, kPoint,&
-            & iAtInCentralRegion, electronicSolver, tDefinedFreeE)
+            & iAtInCentralRegion, electronicSolver, tDefinedFreeE, allocated(halogenXCorrection))
       end if
 
       if (tConverged .or. tStopScc) then
@@ -723,7 +730,7 @@ contains
           & qOutput, q0, skHamCont, skOverCont, pRepCont, neighbourList, nNeighbourSk,&
           & nNeighbourRep, species, img2CentCell, iSparseStart, orb, potential, coord, derivs,&
           & iRhoPrim, thirdOrd, qDepExtPot, chrgForces, dispersion, rangeSep, SSqrReal, over,&
-          & denseDesc, deltaRhoOutSqr, tPoisson)
+          & denseDesc, deltaRhoOutSqr, tPoisson, halogenXCorrection)
 
       if (tCasidaForces) then
         derivs(:,:) = derivs + excitedDerivs
@@ -736,7 +743,7 @@ contains
             & q0, skHamCont, skOverCont, pRepCont, neighbourList, nNeighbourSk, nNeighbourRep,&
             & species, img2CentCell, iSparseStart, orb, potential, coord, latVec,&
             & invLatVec, cellVol, coord0, totalStress, totalLatDeriv, intPressure, iRhoPrim,&
-            & dispersion)
+            & dispersion, halogenXCorrection)
         call env%globalTimer%stopTimer(globalTimers%stressCalc)
         call printVolume(cellVol)
 
@@ -3438,8 +3445,8 @@ contains
 
     energy%atomElec(:) = energy%atomNonSCC + energy%atomSCC + energy%atomSpin + energy%atomDftbu&
         & + energy%atomLS + energy%atomExt + energy%atom3rd + energy%atomOnSite
-    energy%atomTotal(:) = energy%atomElec + energy%atomRep + energy%atomDisp
-    energy%Etotal = energy%Eelec + energy%Erep + energy%eDisp
+    energy%atomTotal(:) = energy%atomElec + energy%atomRep + energy%atomDisp + energy%atomHalogenX
+    energy%Etotal = energy%Eelec + energy%Erep + energy%eDisp + energy%eHalogenX
     energy%EMermin = energy%Etotal - sum(TS)
     ! extrapolated to 0 K
     energy%Ezero = energy%Etotal - 0.5_dp * sum(TS)
@@ -5006,7 +5013,7 @@ contains
       & qOutput, q0, skHamCont, skOverCont, pRepCont, neighbourList, nNeighbourSK, nNeighbourRep,&
       & species, img2CentCell, iSparseStart, orb, potential, coord, derivs, iRhoPrim, thirdOrd,&
       & qDepExtPot, chrgForces, dispersion, rangeSep, SSqrReal, over, denseDesc, deltaRhoOutSqr,&
-      & tPoisson)
+      & tPoisson, halogenXCorrection)
 
     !> Environment settings
     type(TEnvironment), intent(in) :: env
@@ -5110,6 +5117,9 @@ contains
     !> whether Poisson solver is used
     logical, intent(in) :: tPoisson
 
+    !> Correction for halogen bonds
+    type(THalogenX), allocatable, intent(inout) :: halogenXCorrection
+
     ! Locals
     real(dp), allocatable :: tmpDerivs(:,:)
     real(dp), allocatable :: dummyArray(:,:)
@@ -5204,6 +5214,10 @@ contains
       call dispersion%addGradients(derivs)
     end if
 
+    if (allocated(halogenXCorrection)) then
+      call halogenXCorrection%addGradients(derivs, coord, species, neighbourList, img2CentCell)
+    end if
+
     if (allocated(rangeSep)) then
       call unpackHS(SSqrReal, over, neighbourList%iNeighbour, nNeighbourSK, denseDesc%iAtomStart,&
           & iSparseStart, img2CentCell)
@@ -5227,7 +5241,7 @@ contains
   subroutine getStress(env, sccCalc, thirdOrd, tExtField, nonSccDeriv, rhoPrim, ERhoPrim, qOutput,&
       & q0, skHamCont, skOverCont, pRepCont, neighbourList, nNeighbourSk, nNeighbourRep, species,&
       & img2CentCell, iSparseStart, orb, potential, coord, latVec, invLatVec, cellVol, coord0,&
-      & totalStress, totalLatDeriv, intPressure, iRhoPrim, dispersion)
+      & totalStress, totalLatDeriv, intPressure, iRhoPrim, dispersion, halogenXCorrection)
 
     !> Environment settings
     type(TEnvironment), intent(in) :: env
@@ -5319,6 +5333,9 @@ contains
     !> dispersion interactions
     class(DispersionIface), allocatable, intent(inout) :: dispersion
 
+    !> Correction for halogen bonds
+    type(THalogenX), allocatable, intent(inout) :: halogenXCorrection
+
     real(dp) :: tmpStress(3, 3)
     logical :: tImHam
 
@@ -5352,6 +5369,12 @@ contains
 
     if (allocated(dispersion)) then
       call dispersion%getStress(tmpStress)
+      totalStress(:,:) = totalStress + tmpStress
+    end if
+
+    if (allocated(halogenXCorrection)) then
+      call halogenXCorrection%getStress(tmpStress, coord, neighbourList, species, img2CentCell,&
+          & cellVol)
       totalStress(:,:) = totalStress + tmpStress
     end if